fix bug leading to a ClassCastException when trying to create any type of project with idea
we the build a different flavor of the ide in many class loader are involed. The id of the rego module type was not correspoding to the one defined in the class. Consequently it was raising this exception when we try to create a project with idea

2020-08-13 19:17:12,069 [   9667]  ERROR - llij.ide.plugins.PluginManager - class com.intellij.openapi.module.UnknownModuleType cannot be cast to class org.openpolicyagent.ideaplugin.ide.extensions.RegoModuleType (com.intellij.openapi.module.UnknownModuleType is in unnamed module of loader com.intellij.util.lang.UrlClassLoader @4f933fd1; org.openpolicyagent.ideaplugin.ide.extensions.RegoModuleType is in unnamed module of loader com.intellij.ide.plugins.cl.PluginClassLoader @1b14e496)
java.lang.ClassCastException: class com.intellij.openapi.module.UnknownModuleType cannot be cast to class org.openpolicyagent.ideaplugin.ide.extensions.RegoModuleType (com.intellij.openapi.module.UnknownModuleType is in unnamed module of loader com.intellij.util.lang.UrlClassLoader @4f933fd1; org.openpolicyagent.ideaplugin.ide.extensions.RegoModuleType is in unnamed module of loader com.intellij.ide.plugins.cl.PluginClassLoader @1b14e496)

Signed-off-by: Vincent Gramer <vgramer@gmail.com>