chore: Update to OPA 1.0+
I am interested in doing some more work on this project, updating to OPA
1.0 is the first issue I ran into.

Because of https://github.com/open-policy-agent/opa/pull/7410, I have
updated the formatting tests to use v1 Rego.

Much of the current testing largely testing OPA's own functionality, but
I've tried to leave it mostly as is for now to keep things small and
easier to review.

Signed-off-by: Charlie Egan <charlieegan3@users.noreply.github.com>