chore: Update to OPA 1.0+ (#184)
* chore: Update to OPA 1.0+

I am interested in doing some more work on this project, updating to OPA
1.0 is the first issue I ran into.

Because of https://github.com/open-policy-agent/opa/pull/7410, I have
updated the formatting tests to use v1 Rego.

Much of the current testing largely testing OPA's own functionality, but
I've tried to leave it mostly as is for now to keep things small and
easier to review.

Signed-off-by: Charlie Egan <charlieegan3@users.noreply.github.com>

* ci: Use a supported cache version

 This request has been automatically failed because it uses a deprecated
 version of `actions/cache: v2`. Please update your workflow to use
 v3/v4 of actions/cache to avoid interruptions. Learn more:
 https://github.blog/changelog/2024-12-05-notice-of-upcoming-releases-and-breaking-changes-for-github-actions/#actions-cache-v1-v2-and-actions-toolkit-cache-package-closing-down

Signed-off-by: Charlie Egan <charlieegan3@users.noreply.github.com>

---------

Signed-off-by: Charlie Egan <charlieegan3@users.noreply.github.com>
Co-authored-by: Charlie Egan <charlieegan3@users.noreply.github.com>